Nurk on double, mitte int.

+++ src/src/CalculateDegrees.java
@@ -3,7 +3,7 @@ import java.io.IOException;
 import static java.lang.Math.abs;
 
 public class CalculateDegrees {
-    private int difference;
+    private double difference;
     private int hour;
     private int minute;
 
@@ -16,17 +16,17 @@ public class CalculateDegrees {
         if (hour < 0 || hour > 12 || minute < 0 || minute > 60) {
             throw new IOException("Tunnid või minutid on valed.");
         } else {
-            int hourDegrees = 360 / 12 * hour;
-            int minutesDegrees = 360 / 60 * minute;
+            double hourDegrees = 360 / 12 * hour;
+            double minutesDegrees = 360 / 60 * minute;
             // The minute hand moves if minute hand moves.
             // Maximum movement for hour is 30 degrees and minimum 0.
-            int hourDegreesMovement = minutesDegrees / 12;
+            double hourDegreesMovement = minutesDegrees / 12;
             this.difference = abs(abs(hourDegrees - minutesDegrees) - hourDegreesMovement);
         }
 
     }
 
-    int getDifference() {
+    double getDifference() {
         return difference;
     }
